+++ src/main/kotlin/com/mituuz/fuzzier/components/FuzzyComponent.kt
@@ -26,7 +26,6 @@ package com.mituuz.fuzzier.components
 import com.intellij.ui.EditorTextField
 import com.intellij.ui.components.JBList
 import com.mituuz.fuzzier.entities.FuzzyContainer
-import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import javax.swing.JPanel
 
 open class FuzzyComponent : JPanel() {