+++ src/main/kotlin/com/mituuz/fuzzier/entities/ScoreCalculator.kt
@@ -25,7 +25,7 @@ package com.mituuz.fuzzier.entities
 
 import com.intellij.openapi.components.service
 import com.mituuz.fuzzier.entities.FuzzyMatchContainer.FuzzyScore
-import com.mituuz.fuzzier.settings.FuzzierSettingsService
+import com.mituuz.fuzzier.settings.FuzzierGlobalSettingsService
 import org.apache.commons.lang3.StringUtils
 
 class ScoreCalculator(searchString: String) {
@@ -40,14 +40,13 @@ class ScoreCalculator(searchString: String) {
     private var filenameIndex: Int = 0
 
     // Set up match settings
-    private val settings = service<FuzzierSettingsService>().state
-    private var tolerance = settings.tolerance
-    private var multiMatch = settings.multiMatch
-    private var matchWeightSingleChar = settings.matchWeightSingleChar
-    private var matchWeightStreakModifier = settings.matchWeightStreakModifier
-    private var matchWeightPartialPath = settings.matchWeightPartialPath
-    private var matchWeightFilename = settings.matchWeightFilename
-    private var ignoredCharacters: Set<Char> = settings.ignoredCharacters.toSet()
+    private val globalState = service<FuzzierGlobalSettingsService>().state
+    private var tolerance = globalState.tolerance
+    private var multiMatch = globalState.multiMatch
+    private var matchWeightSingleChar = globalState.matchWeightSingleChar
+    private var matchWeightStreakModifier = globalState.matchWeightStreakModifier
+    private var matchWeightPartialPath = globalState.matchWeightPartialPath
+    private var matchWeightFilename = globalState.matchWeightFilename
 
     var currentFilePath = ""
     private var longestStreak: Int = 0
@@ -128,7 +127,7 @@ class ScoreCalculator(searchString: String) {
 
     private fun processChar(searchStringPartChar: Char): Boolean {
         if (filePathIndex >= currentFilePath.length) {
-            return false;
+            return false
         }
         val filePathPartChar = currentFilePath[filePathIndex]
         if (searchStringPartChar == filePathPartChar) {
@@ -222,9 +221,5 @@ class ScoreCalculator(searchString: String) {
 
     fun setTolerance(value: Int) {
         tolerance = value
-    }
-
-    fun setIgnoredCharacters(characterString: String) {
-        ignoredCharacters = characterString.toSet()
     }
 }