+++ src/main/kotlin/com/mituuz/fuzzier/settings/FuzzierSettingsService.kt
@@ -24,48 +24,28 @@ SOFTWARE.
 package com.mituuz.fuzzier.settings
 
 import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.annotations.OptionTag
 import com.mituuz.fuzzier.entities.FuzzyMatchContainer
-import com.mituuz.fuzzier.entities.FuzzyContainer.FilenameType
-import com.mituuz.fuzzier.entities.FuzzyContainer.FilenameType.FILE_PATH_ONLY
-import com.mituuz.fuzzier.settings.FuzzierSettingsService.RecentFilesMode.RECENT_PROJECT_FILES
 import javax.swing.DefaultListModel
 
 @State(
     name = "com.mituuz.fuzzier.FuzzierSettings",
-    storages = [Storage("FuzzierSettings.xml")]
+    storages = [Storage("FuzzierSettings.xml")],
+    reloadable = true
 )
+@Service(Service.Level.PROJECT)
 class FuzzierSettingsService : PersistentStateComponent<FuzzierSettingsService.State> {
     class State {
         var modules: Map<String, String> = HashMap()
         var isProject = false
-        var recentFilesMode: RecentFilesMode = RECENT_PROJECT_FILES
         @OptionTag(converter = FuzzyMatchContainer.SerializedMatchContainerConverter::class)
         var recentlySearchedFiles: DefaultListModel<FuzzyMatchContainer.SerializedMatchContainer>? = DefaultListModel()
 
-        var splitPosition: Int = 300
         var exclusionSet: Set<String> = setOf("/.idea/*", "/.git/*", "/target/*", "/build/*", "/.gradle/*", "/.run/*")
         var ignoredCharacters: String = ""
-        var newTab: Boolean = false
-        var prioritizeShorterDirPaths = true
-        var debouncePeriod: Int = 80
-        var resetWindow = false
-        var fileListLimit: Int = 50
-
-        var filenameType: FilenameType = FILE_PATH_ONLY
-        var highlightFilename = false
-        var fileListFontSize = 14
-        var previewFontSize = 0
-        var fileListSpacing = 0
-
-        var tolerance = 0
-        var multiMatch = false
-        var matchWeightPartialPath = 10
-        var matchWeightSingleChar = 5
-        var matchWeightStreakModifier = 10
-        var matchWeightFilename = 10
 
         fun getRecentlySearchedFilesAsFuzzyMatchContainer(): DefaultListModel<FuzzyMatchContainer> {
             val list = recentlySearchedFiles ?: DefaultListModel()
@@ -81,11 +61,5 @@ class FuzzierSettingsService : PersistentStateComponent<FuzzierSettingsService.S
 
     override fun loadState(p0: State) {
         this.state = p0
-    }
-
-    enum class RecentFilesMode(val text: String) {
-        NONE("None"),
-        RECENT_PROJECT_FILES("Recent project files"),
-        RECENTLY_SEARCHED_FILES("Recently searched files")
     }
 }