+++ src/test/kotlin/com/mituuz/fuzzier/entities/RowContainerTest.kt
@@ -1,13 +1,13 @@
 package com.mituuz.fuzzier.entities
 
-import com.mituuz.fuzzier.settings.FuzzierSettingsService
+import com.mituuz.fuzzier.settings.FuzzierGlobalSettingsService
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 
 class RowContainerTest {
     @Test
     fun displayString() {
-        val state = FuzzierSettingsService.State()
+        val state = FuzzierGlobalSettingsService.State()
         val container = RowContainer("", "", "filename", 0)
 
         assertEquals("filename:0", container.getDisplayString(state))