+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Version 1.4.1
 - Refactoring match containers
+- Properly separate project and application settings
 
 ## Version 1.4.0
 - Add option to ignore characters from search